# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "golang作品",
        "title": "Gin-Backend",
        "termsOfService": "github",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/user/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Order AddUser",
                "parameters": [
                    {
                        "description": "Add User",
                        "name": "u",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Order DeleteUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": " DeleteUser",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Order EditUser",
                "parameters": [
                    {
                        "description": "Edit User",
                        "name": "u",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Order UserInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Info",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Order UserList",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.User": {
            "type": "object",
            "required": [
                "mobile"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "isLocked": {
                    "type": "boolean"
                },
                "mobile": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}